subaccount and customer ad user creation toggles separated

+++ azure-functions/src/main/java/com/function/TekvLSDeleteCustomerById.java
@@ -79,15 +79,15 @@ public class TekvLSDeleteCustomerById
 			PreparedStatement deleteStmt = connection.prepareStatement(deleteSql)) {
 			context.getLogger().info("Successfully connected to: " + System.getenv("POSTGRESQL_SERVER"));
 
-			if(FeatureToggles.INSTANCE.isFeatureActive("ad-user-creation")) {
+			if (FeatureToggles.INSTANCE.isFeatureActive("ad-subaccount-user-creation") || FeatureToggles.INSTANCE.isFeatureActive("ad-customer-user-creation")) {
 				String getAllEmailsSql = "SELECT sa.subaccount_admin_email,sa.subaccount_id,s.customer_id as subaccount_customer_id,ca.admin_email,ca.customer_id " +
 						"FROM subaccount_admin sa " +
 						"JOIN subaccount s ON sa.subaccount_id = s.id " +
 						"FULL JOIN customer_admin ca ON sa.subaccount_admin_email = ca.admin_email " +
 						"WHERE s.customer_id = ?::uuid OR ca.customer_id = ?::uuid;";
 				try(PreparedStatement getAllEmailsStmt = connection.prepareStatement(getAllEmailsSql)){
-					getAllEmailsStmt.setString(1,id);
-					getAllEmailsStmt.setString(2,id);
+					getAllEmailsStmt.setString(1, id);
+					getAllEmailsStmt.setString(2, id);
 					ResultSet resultSet = getAllEmailsStmt.executeQuery();
 					while (resultSet.next()){
 						String customerId =  resultSet.getString("customer_id");
@@ -95,26 +95,27 @@ public class TekvLSDeleteCustomerById
 						String subaccountCustomerId =  resultSet.getString("subaccount_customer_id");
 						String subaccountAdminEmail = resultSet.getString("subaccount_admin_email");
 
-						if(subaccountCustomerId == null){
-							if(FeatureToggles.INSTANCE.isFeatureActive("ad-customer-user-creation"))
+						if (subaccountCustomerId == null || (!FeatureToggles.INSTANCE.isFeatureActive("ad-subaccount-user-creation") && adminEmail != null)) {
+							if (FeatureToggles.INSTANCE.isFeatureActive("ad-customer-user-creation")) {
 								GraphAPIClient.deleteGuestUser(adminEmail, context);
-								context.getLogger().info("Guest User deleted successfully from Active Directory (email: "+adminEmail+").");
+								context.getLogger().info("Guest User deleted successfully from Active Directory (email: " + adminEmail + ").");
+							}
 							continue;
 						}
 
-						if(subaccountCustomerId.equals(id)){
-							if(adminEmail==null || subaccountCustomerId.equals(customerId)){
+						if (subaccountCustomerId.equals(id) && FeatureToggles.INSTANCE.isFeatureActive("ad-subaccount-user-creation")) {
+							if (adminEmail == null || subaccountCustomerId.equals(customerId)) {
 								GraphAPIClient.deleteGuestUser(subaccountAdminEmail, context);
-								context.getLogger().info("Guest User deleted successfully from Active Directory (email: "+subaccountAdminEmail+").");
+								context.getLogger().info("Guest User deleted successfully from Active Directory (email: " + subaccountAdminEmail + ").");
 								continue;
 							}
-							GraphAPIClient.removeRole(subaccountAdminEmail,SUBACCOUNT_ADMIN,context);
-							context.getLogger().info("Guest User Role (Subaccount Admin) removed successfully from Active Directory (email: "+subaccountAdminEmail+").");
+							GraphAPIClient.removeRole(subaccountAdminEmail, SUBACCOUNT_ADMIN, context);
+							context.getLogger().info("Guest User Role (Subaccount Admin) removed successfully from Active Directory (email: " + subaccountAdminEmail + ").");
 							continue;
 						}
 
-						if(FeatureToggles.INSTANCE.isFeatureActive("ad-customer-user-creation")) {
-							GraphAPIClient.removeRole(adminEmail,CUSTOMER_FULL_ADMIN,context);
+						if (FeatureToggles.INSTANCE.isFeatureActive("ad-customer-user-creation")) {
+							GraphAPIClient.removeRole(adminEmail, CUSTOMER_FULL_ADMIN, context);
 							context.getLogger().info("Guest User Role (Customer Admin) removed successfully from Active Directory (email: " + adminEmail + ").");
 						}
 					}
